Create HierarchialInheritance.java
The program code has two child classes Hockey Player and Cricket player extended from the parent class Player. The program uses constructors and method overriding to efficiently display the no. of goals for hockey player and no. of wickets and runs for cricket player. The program uses switch case in the main function to choose between cricket or hockey player